requirements.txt
torch
opencv-python
numpy
pillow
matplotlib
scipy
pandas
simple-pid
tk
requests
ultralytics 